handle NumPy style docstring params containing a colon #189
The problematic case was:

    name : str
        name, default: World

The second line wasn't treated as a description if it contained a colon
and ended up ignored

We parse line by line, and used to keep track only of the next line.
I added tracking of the previous line so that I can compare the indentations.
That allows to distinguish descriptions and a parameter definitions.